🪺 Update for Unicode 14
Unicode 14 is here, and with it are some breaking changes in `emoji-regex`, which mean we need to change how we generate our own stuff.

To accomplish this, we're moving to the slightly-more-upstream package, `emoji-test-regex-pattern`, and using its `java` regex variation.

This also means removing all but the one regular expression. We previously marked `Emoji` as deprecated, but the plan now is to mark `Text` as deprecated also.